% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdd_mcmc.R
\name{bcdd_mcmc}
\alias{bcdd_mcmc}
\title{The MCMC algorithm for Constrained Bayes density deconvolution estimator.}
\usage{
bcdd_mcmc(w, sd_u, n.burnin = 1000, n.MCMC = 5000, hpar = list(m = 20, K =
  8, Xi_1 = 1, Xi_2 = 4, lambda = 2, shape.lambda = 2, tt = 2.5))
}
\arguments{
\item{w}{A vector containing the proxy variables.}

\item{sd_u}{A vector containing the standard deviation of error, if the error
is homoscedastic, then it should be a scalar.}

\item{n.burnin}{A number (default 1000) specifies the number of burning steps.}

\item{n.MCMC}{A number (default 5000) specifies the number of total MCMC steps.}

\item{hpar}{A list containing the hyperparameters in the algorithm, the 
default values are chosen if left empty. m is the concentration parameter and 
K is the number of categories of Dirichlet distribution prior for p. Xi_1, Xi_2 
are the shape and rate parameters of the Gamma prior for beta. lambda is the 
rate parameter of the exponential prior for alpha. shape.lambda is one parameter
in the proposal distribution for the MH step of alpha. tt corresponds to the 
lower bound that is prespecified for alpha.}
}
\value{
A list containing the expected posterior density of the latent true 
variable (y) evaluated on a grid (x) chosen automatically in the function.
}
\description{
The main function \code{bcdd_mcmc} in this package can be used to deconvolve 
a density with a unimodal and symmetric shape. The method is coined the name 
Constrained Bayes and is essentially a nonparametric approach using Bayesian
hierarchical models. The speed is bumped by utilizing functions written in Rcpp.
}
\examples{
df.t <- 5
x <- rt(n, df = df.t)
sd_u <- 1.29
u <- rnorm(n, mean = 0, sd = sd_u)               
w <- x + u
ddsc_mcmc(w, sd_u, n.burnin=100, n.MCMC=500)
}
