% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myfunction.R
\name{fun_integ}
\alias{fun_integ}
\title{Calculate the integrand for the marginal distribution of X over theta.}
\usage{
fun_integ(x, pi, shape, rate)
}
\arguments{
\item{x}{a number (vector) specifies the value of the domain argument.}

\item{pi}{a vector containing the component probabilities of the Gammas}

\item{shape, }{rate numeric vectors corresponding to shape and rate parameters
of Gammas}
}
\value{
The (posterior) marginal distribution of X (over theta) at the input 
parameter x.
}
\description{
The (posterior) marginal distribution of X equals the integral of the conditional 
distribution of X given theta, uniform over [-theta, theta], weighted by the 
(posterior) distribution of theta, a mixture of Gammas. The function fun_integ
explicitly computes the integrand as a function of theta.
}
\examples{
pi <- c(0.2,0.5,0.3)
shape <- c(2.3,3.3,4.3) 
rate <- c(1.5,2.5,3.5)
theta <- 0.1 
fun_integ(theta, pi, shape, rate)
theta <- c(0.1,0.5,3)
fun_integ(theta, pi, shape, rate)
}
