% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mydgamma}
\alias{mydgamma}
\title{mydgamma}
\usage{
mydgamma(x, s, r)
}
\arguments{
\item{x}{a vector of values at which the gamma densities are evaluated.}

\item{s}{a vector of shape parameters.}

\item{r}{a vector of rate parameters.}
}
\value{
A matrix with (i,j) component equals the j-th gamma density evaluated
at x_i.
}
\description{
The function \code{mydgamma} is meant for computing the matrix
as the input for the function \code{mysample} where the family of densities 
are gamma. See also \code{mysample}.
}
\examples{
N <- 500
K <- 3
alpha <- c(2.928, 2.979, 2.754)
beta <- c(0.124, 0.238, 0.045)
theta <-  rgamma(N, shape=2.6, rate=0.3)
d.ordinates <- mydgamma(theta, alpha, beta)
}
