% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mysample}
\alias{mysample}
\title{mysample}
\usage{
mysample(K, ord, pvec)
}
\arguments{
\item{K}{a value that gives the total category that z varies from.}

\item{ord}{a matrix with (i,j) component equals the prespecified density evaluated
at the latent variable coexisting with z_i.}

\item{pvec}{a vector of probabilities that is associated with the (latent) 
true prior of these component probability.}
}
\value{
A vector containing the posterior samples z_1,...,z_n.
}
\description{
The function \code{mysample} is used to update/sample the indicators 
z_1, ..., z_n. The core function is the basic sample function. The component 
probability equals a density evaluated at the latent variable coexisting 
with z and a probability that is associated with the (latent) true prior of 
these component probability.
}
\examples{
set.seed(123)
N <- 500
K <- 3
alpha <- c(2.928, 2.979, 2.754)
beta <- c(0.124, 0.238, 0.045)
p <- c(0.069, 0.758, 0.173)
theta <-  rgamma(N, shape=2.6, rate=0.3)
d.ordinates <- matrix(0, nrow=N, ncol=K)
for(jj in 1:K){
    d.ordinates[,jj] <- dgamma(theta, shape=alpha[jj], rate=beta[jj])
}
z <- mysample(1:K, d.ordinates, p) 
}
